move matcher into its own file
